Replace Exception-oriented approach for flow control by ServiceResponse
ServiceResponse is an interface that represents the concepts of a response from a Service. The classes that implement it can be returned by Service methods. The defined implementations are ServiceStatus, ServiceEntity and ServiceError. ServicePrimitive is a generic subclass of ServiceEntity that wraps a primitive type (that cannot be monkey-patched to implement ServiceResponse).

This approach is not type-safe, so ServiceResponse interface MUST be used carefully to avoid bugs. It could be converted into a generic class to ensure type-safety, but the current approach is promising in order to enhance code readability as long as it's accompanied with proper documentation.